requirements.txt
python-telegram-bot[job-queue]==21.6
aiohttp==3.10.11
beautifulsoup4==4.12.2
python-dotenv==1.0.0
playwright==1.49.0
Brotli==1.1.0
